Moved the auth stuff into its own app
