--- python/aristotle-mdr-api/aristotle_mdr_api/admin.py
@@ -1,4 +1,0 @@
-from django.contrib import admin
-from aristotle_mdr_api.models import AristotleToken
-
-admin.site.register(AristotleToken)
