+++ python/aristotle-mdr-api/aristotle_mdr_api/settings.py
@@ -1,5 +1,6 @@
 REQUIRED_APPS = (
     'aristotle_mdr_api',
+    'aristotle_mdr_api.token_auth',
     'rest_framework',
     'rest_framework_swagger',
     'django_filters',
@@ -12,7 +13,7 @@ REST_FRAMEWORK = {
         'rest_framework.permissions.IsAuthenticated',
     ),
     'DEFAULT_AUTHENTICATION_CLASSES': (
-        'aristotle_mdr_api.authentication.AristotleTokenAuthentication',
+        'aristotle_mdr_api.token_auth.authentication.AristotleTokenAuthentication',
         'rest_framework.authentication.SessionAuthentication',
     ),
     # 'DEFAULT_VERSIONING_CLASS': 'rest_framework.versioning.NamespaceVersioning'
