+++ python/aristotle-mdr-api/aristotle_mdr_api/token_auth/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'aristotle_mdr_api.token_auth.apps.TokenAuthConfig'
