+++ python/aristotle-mdr-api/aristotle_mdr_api/token_auth/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+class TokenAuthConfig(AppConfig):
+    name = 'aristotle_mdr_api.token_auth'
+    label = 'aristotle_mdr_tokenauth'
+    verbose_name = 'Token Authentication'
