+++ python/aristotle-mdr-api/aristotle_mdr_api/token_auth/authentication.py
@@ -1,5 +1,5 @@
 from rest_framework.authentication import TokenAuthentication
-from aristotle_mdr_api.models import AristotleToken
+from aristotle_mdr_api.token_auth.models import AristotleToken
 
 class AristotleTokenAuthentication(TokenAuthentication):
     model = AristotleToken
